Move parameters_test.py to new folder tests/core_tests
